Novas configurações para o deploy

+++ byustechnology/gabineteagil/src/views/assunto/index.blade.php
@@ -20,36 +20,44 @@
     <a href="#" data-toggle="modal" data-target="#m-search" class="btn btn-sm btn-primary"><i class="fas fa-search fa-fw mr-1"></i> Buscar</a>
     {!! request()->query() ? '<a href="' . url(request()->url()) . '" class="btn btn-sm btn-link text-danger"><i class="far fa-times-circle mr-2"></i>Cancelar filtro</a>' : null !!}
 
-    <div class="table-responsive mt-3">
-        <table class="table">
-            <thead>
-                <tr>
-                    <th>Título</th>
-                    <th>Cor</th>
-                    <th class="text-right">Nº Ocorrências</th>
-                    <th class="table-actions">Ações</th>
-                </tr>
-            </thead>
-            <tbody>
-                @foreach($assuntos as $assunto)
-                <tr>
-                    <td>
-                        <a href="{{ url($assunto->path()) }}"><strong>{{ $assunto->titulo }}</strong></a><br>
-                        <small class="text-muted">{{ Str::limit($assunto->descricao, 100) }}</small>
-                    </td>
-                    <td><i data-toggle="tooltip" title="{{ $assunto->cor }}" class="fas fa-square fa-fw mr-1 shadow-sm" style="color: {{ $assunto->cor }}"></i></td>
-                    <td class="text-right">{{ $assunto->ocorrencias_count }}</td>
-                    <td class="table-actions">
-                        {!! Form::open(['url' => route('assunto.destroy', ['assunto' => $assunto->id]), 'method' => 'delete']) !!}
-                        <a data-toggle="tooltip" title="Editar" href="{{ route('assunto.edit', ['assunto' => $assunto]) }}" class="btn btn-table-actions btn-link"><i class="far fa-edit fa-fw"></i></a>
-                        <button data-toggle="tooltip" title="Remover" type="sumbit" class="btn btn-table-actions text-danger btn-link"><i class="far fa-trash-alt fa-fw"></i></button>
-                        {!! Form::close() !!}
-                    </td>
-                </tr>
-                @endforeach
-            </tbody>
-        </table>
-    </div>
+    @component('gabinete::components.card')
+        @if($assuntos->count())
+            <div class="table-responsive mt-3">
+                <table class="table">
+                    <thead>
+                        <tr>
+                            <th>Título</th>
+                            <th>Cor</th>
+                            <th class="text-right">Nº Ocorrências</th>
+                            <th class="table-actions">Ações</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        @foreach($assuntos as $assunto)
+                        <tr>
+                            <td>
+                                <a href="{{ url($assunto->path()) }}"><strong>{{ $assunto->titulo }}</strong></a><br>
+                                <small class="text-muted">{{ Str::limit($assunto->descricao, 100) }}</small>
+                            </td>
+                            <td><i data-toggle="tooltip" title="{{ $assunto->cor }}" class="fas fa-square fa-fw mr-1 shadow-sm" style="color: {{ $assunto->cor }}"></i></td>
+                            <td class="text-right">{{ $assunto->ocorrencias_count }}</td>
+                            <td class="table-actions">
+                                {!! Form::open(['url' => route('assunto.destroy', ['assunto' => $assunto->id]), 'method' => 'delete']) !!}
+                                <a data-toggle="tooltip" title="Editar" href="{{ route('assunto.edit', ['assunto' => $assunto]) }}" class="btn btn-table-actions btn-link"><i class="far fa-edit fa-fw"></i></a>
+                                <button data-toggle="tooltip" title="Remover" type="sumbit" class="btn btn-table-actions text-danger btn-link"><i class="far fa-trash-alt fa-fw"></i></button>
+                                {!! Form::close() !!}
+                            </td>
+                        </tr>
+                        @endforeach
+                    </tbody>
+                </table>
+            </div>
+
+            {!! $assuntos->links() !!}
+        @else
+            @include('gabinete::components.no-results')
+        @endif
+    @endcomponent
 </div>
 
 @include('gabinete::assunto.partials.search')
